add initial CMakeLists, fix to run under linux

+++ RayTracerOne/vec3.h
@@ -59,7 +59,7 @@ inline std::ostream& operator << (std::ostream &os, const vec3 &v) {
 }
 
 inline void vec3::make_unit_vector() {
-	float k = 1.0f / std::sqrtf(e[0] * e[0] + e[1] * e[1] + e[2] * e[2]);
+	float k = 1.0f / std::sqrt(e[0] * e[0] + e[1] * e[1] + e[2] * e[2]);
 	e[0] *= k; e[1] *= k; e[2] *= k;
 }
